Split owner commands into individual files
