+++ commands/Owner/updateAlternion.js
@@ -1,0 +1,11 @@
+const alt = require("./alternion");
+
+module.exports = {
+	name: "do",
+	args: [1,200],
+	help: "Runs provided text as code",
+	users: ["337541914687569920"],
+	execute: async (message,args) => {
+		alt.globalJsonUpdate();
+	}
+}
